+++ src/mcmu/__init__.py
@@ -8,8 +8,7 @@ try:
 except PackageNotFoundError:
     __version__ = "Version  Not Found"
 
-USER_AGENT = "Josiah-Jarvis/MCMU/"
-USER_AGENT += f"{__version__} "
+USER_AGENT = f"Josiah-Jarvis/MCMU/{__version__} "
 USER_AGENT += "(https://github.com/Josiah-Jarvis/MCMU)"
 
 logger = getLogger(__name__)